➖ Remove unused prompts and location memo extraction code, add new parsing agent functionality.
