--- app/langgraph/parsing_agent/prompts.py
@@ -1,13 +1,0 @@
-title_prompt = (
-    "아래 일정을 반드시 한 줄로 요약해줘. "
-    "- 시간, 날짜, 장소, 사람 이름, 준비물 등은 절대 포함하지 마. "
-    "- 반드시 구체적인 행동 또는 주제만 남겨. "
-    "- '일정', '없음', '모름', '기타' 등 의미 없는 단어는 절대 사용하지 마. "
-    "- 예시처럼 핵심 행동/주제만 남기고, 불명확하면 원문에서 가장 중요한 동작/행동을 뽑아. "
-    "예시) '로건이와 산책', '김부장 미팅', '삼겹살 회식', '짬뽕 식사', '팀 회의', '보고서 작성', '수업 듣기', '운동하기', '친구와 영화관람', '가족 모임', '병원 진료', '쇼핑', '스터디 참여', '프로젝트 회의', '면접', '출장', '강의 준비', '책 읽기', '청소', '요리', '반려동물 산책'"
-)
-
-memo_prompt = (
-    "아래 일정에서 준비물, 신경쓸 것, 참고사항, 감정표현, 정보 등을 한 줄로 요약해줘. "
-    "예시) '피푸봉투 챙기기', '메로나 지참', '우산 챙기기'"
-)
