+++ app/utils/prompt/parse_prompts.py
@@ -1,0 +1,11 @@
+title_prompt = (
+    "아래 일정을 한 줄로 요약해줘. 시간, 날짜, 장소, 사람 이름, 준비물 등은 가급적 빼고, 핵심 행동/주제만 남겨줘. "
+    "예시처럼 2~10자 이내로 짧게 써줘. 너무 애매하면 원문에서 가장 중요한 단어나 행동을 뽑아도 좋아. "
+    "'일정', '없음', '모름', '기타' 등 의미 없는 단어는 사용하지 마. "
+    "예시) '로건이와 산책', '김부장 미팅', '삼겹살 회식', '짬뽕 식사', '팀 회의', '보고서 작성', '수업 듣기', '운동하기', '친구와 영화관람', '가족 모임', '병원 진료', '쇼핑', '스터디 참여', '프로젝트 회의', '면접', '출장', '강의 준비', '책 읽기', '청소', '요리', '반려동물 산책', '치맥파티', '생일파티', '동호회 모임', '동아리 회식', '스터디 발표', '워크숍', '멘토링', '동문회', '송년회', '신년회', 'MT', '워크아웃', '팀 점심', '팀 저녁', '친구 모임', '가족 식사', '부서 회의', '면담', '상담', '면접', '발표', '강연', '특강', '공연 관람', '전시 관람', '봉사활동', '운동회', '야유회', '캠핑', '등산', '여행', '피크닉', '바베큐', '홈파티'"
+)
+
+memo_prompt = (
+    "아래 일정에서 준비물, 신경쓸 것, 참고사항, 감정표현, 정보 등을 한 줄로 요약해줘. "
+    "예시) '피푸봉투 챙기기', '메로나 지참', '우산 챙기기'"
+)
